requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.14.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
joblib>=1.3.0
protobuf>=4.24.4
